refactor typings, installed numpy
