+++ pyproject.toml
@@ -10,6 +10,7 @@ packages = [{include = "project_pygames"}]
 [tool.poetry.dependencies]
 python = "^3.11"
 pygame = "^2.3.0"
+numpy = "^1.24.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.3.0"
